+++ editor_patch/main.cpp
@@ -629,6 +629,12 @@ void apply_af_level_editor_changes()
     LoadSaveLevel_patch2.install();
     disable_splash_screen_on_load_level.install();
 
+    // Write version 300 to saved group files, instead of 200 (default)
+    // Fixes issue loading events with orientation values (in event.cpp DedEvent__exchange_patch)
+    // Note: Doesn't need to be updated when MAXIMUM_RFL_VERSION is incremented
+    AsmWriter(0x0041D0AF).push(0x12C);
+    AsmWriter(0x0041D0BC).push(0x12C);
+
     // Remove legacy geometry maximums from build output window
     static char new_faces_string[] = "Faces: %d\n";                                  // Replace "Faces: %d/%d\n"
     static char new_face_vertices_string[] = "Face Vertices: %d\n";                  // Replace "Face Vertices: %d/%d\n"