+++ docs/CHANGELOG.md
@@ -130,6 +130,7 @@ Version 1.1.0 (Tamarack): Not yet released
 - Fix "Time Left" text label not being localized
 - Fix `Player Entity Type`, `Player Undercover Suit Entity Type`, `Player Undercover Scientist Entity Type` not always applying
 - Fix lava and acid damage sometimes being attributed to a random player in the game
+- Fix level editor not handling new events with orient values from loaded groups (bump version field for rfg files)
 
 Version 1.0.0 (Maple): Released Jan-25-2025
 --------------------------------